Added example of apache reverse proxy with docker
- The important part is to use the name given with docker-compose

// Dockerfile example:
services:
  hello-node:
    ...
  goodbye-node:
    ...
  web-node:
    ...

The services above are added to the same network so that each service can communicate with one another. In order to get the address of the services within the network you must do the following:

// 000-default.conf example:
ProxyPass /hello http://hello-node:3000/

In the above example, the apache server (web-node) is calling on docker's network and asking for the location of the (hello-node) and getting that node's port 3000. Then apache is mapping that (hello-node) location to localhost/hello